library: fix conversion warnings
- Change the data type of variable 'size' in DltContextData struct
  (dlt_user.h.in). Variable 'size' was used repeated with function
  'sizeof()' many time, so change type to 'size_t' can avoid a lot of
  warnings.

- Adding temporary variable to calculate the datasize header (define in
  dlt_common.h) in 'dlt_common.c' file.

+++ src/lib/dlt_env_ll.c
@@ -125,7 +125,7 @@ int dlt_env_extract_symbolic_ll(char **const env, int8_t *ll)
     if (!(*env))
         return -1;
 
-    if (dlt_env_helper_to_lower(env, &result[0], sizeof(result)) == 0) {
+    if (dlt_env_helper_to_lower(env, &result[0], (int) sizeof(result)) == 0) {
         if (strncmp("default", result, sizeof(result)) == 0)
             *ll = -1;
         else if (strncmp("off", result, sizeof(result)) == 0)
@@ -203,7 +203,7 @@ int dlt_env_extract_ll(char **const env, int8_t *ll)
     }
     else {
         if ((**env >= '0') && (**env < '7')) {
-            *ll = **env - '0';
+            *ll = (int8_t) (**env - '0');
             (*env)++;
         }
         else if (dlt_env_extract_symbolic_ll(env, ll) != 0)
